+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # wavegen
 
+## 0.4.0
+
+- `Waveform` has a parametrized precision now. By default `f32`, can be set to `f64` if needed.
+- Renamed `Waveform::get_sample_rate` to `Waveform::sample_rate`.
+- Removed `Waveform::get_components_len`.
+- Added `Waveform::components` getter.
+
 ## 0.3.0
 
 - `wf` helper macro added.